refactor: move unit.dart into new unit folder
