+++ lib/data/services/local_data_service.dart
@@ -6,7 +6,7 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 import '../../domain/models/event.dart';
 import '../../domain/models/noc.dart';
-import '../../domain/models/unit.dart';
+import '../../domain/models/unit/unit.dart';
 import '../../domain/models/user/user.dart';
 
 class LocalDataService {
